chore: reorganize models
