+++ src/DynamicTreeDataGrid/Models/IDynamicTreeDataGridSource`1.cs
@@ -1,6 +1,6 @@
 ﻿using Avalonia.Controls;
 
-namespace DynamicTreeDataGrid.Source;
+namespace DynamicTreeDataGrid.Models;
 
 public interface IDynamicTreeDataGridSource<TModel> : ITreeDataGridSource<TModel>, IDynamicTreeDataGridSource {
     // TODO: Add filter, state, maybe sort?
