+++ src/DynamicTreeDataGrid/Models/State/GridState.cs
@@ -1,4 +1,4 @@
-﻿namespace DynamicTreeDataGrid.State;
+﻿namespace DynamicTreeDataGrid.Models.State;
 
 public class GridState {
     public IList<ColumnState> ColumnStates { get; set; }
